New fit model for chib to Y(1S) gamma decays.
For chib to Y(1S) gamma fit model:
(1) as was studied there are no scaling needed for chib(1P) signal in comparison with MC (before scaling factor was 1.16)
(2) No difference between widths of chib1(1P) and chib2(1P) signals
(3) Was changed polynomial order for  background component for decays to Y(1S) (Table 3.4)
(4) Difference beetween masses of chib2(3P) and chib1(3P) is fixed to 10.5 MeV (for consistency with converted photons study). In the previous note for decays Y(2S) and Y(3S) this value was already used.

Due to changes in the model for chib to Y(1S) gamma  decays was observed decreasing of chib yields up to 10% for the lower pt bins.